pyproject.toml
[tool.poetry]
name = "dtw-test"
version = "0.1.0"
description = ""
authors = ["hjmr <2620990+hjmr@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dtw-python = "^1.1.10"
matplotlib = "^3.4.3"
h5py = "^3.4.0"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119